js/history: do not remove special params
We used to do this to blacklist certain parameters to not trigger
heavvy actions twice when navigating through history. We also do
never want to see such parameters in our address bar.

This conflicts with our slightly more complicated new filter URLs,
unless we are able to handle them on client side this is disabled
for now. IMO it should be handled server-side and eventually invol-
ving one more redirect.