Ensure that the ini writer deletes properties properly
The fix for #5958 introduced with commit 6aaff6b3 caused the property
deletion of the PreservingIniWriter be even more faulty than before. The
actual fix for this issue is an improved procedure for how properties are
deleted as the initial implementation suffered from a few
misunderstandings in how properties in Zend_Ini files need to be handled.

fixes #6599
fixes #6601
fixes #6600
fixes #6598
fixes #6614