Add web setup
Page config and dependency handling is implemented and works like a
charm. Though there is one known bug that occurs when navigating back,
taking changes and moving onward to a page that needs to update its
state due to the changes while adding some elements that have similar
names as the ones shown before. This is only an issue with hidden elements.

refs #7163