Add CLI command to create the configuration directory
It is now not necessary anymore to generate the token as the web server's
user. A user now creates the configuration directory using the CLI as well,
while the SGID is set (in the default mode) causing the token only to be
accessible by the super user who owns the directory/file and the web
server's group.

refs #7163
refs #7408
refs #7410