Chart/InlinePie: smaller, white border for hover
I roughly changed the piechart defaults to their original (icon-like)
size. However, this is NOT how we should handle such changes. We MUST
be able to style them via CSS and get rid of that incredible amount of
HTML attributes.

I want to be able to set the height directly in CSS as em/px. And I want
also be able to set it to 100% and size the outer container.

refs #7077