Revert "Use /usr/bin/env instead of hard-coding paths"
This reverts commit 42b0469a96a17fc2f0eac200b94f46990008ee60.

Packaged verions of our scripts must use a system installation of the interpreter for obvious reasons.
I also disapprove using env in our Git. We want to use a system installation of the interpreter too. Not something that's supposed to be the interpreter.
Also, potential incomplete or defective local versions of the interpreter must not be our concern.
That installations from source using /usr/bin/php do not work out of the box on FreeBSD for example is of course an issue. But that should not be fixed by blindly using env.

refs #7789