monitoring: Deprecate DataView::addFilter() and DataView::setFilter()
The from now on deprecated interface Filterable has proven that it sucks in the DataView.
Because of requiring us to implement trillion stupid methods, only DataView::applyFilter()
does not forget to handle column validation. Thus only DataView::applyFilter() must be used in order
to apply filters.
For setFilter() a wrapping Filter::matchAny() for the IdoQuery (or the DbQuery or the SimpleQuery I didn't have a look)
is required for the filter to work properly.
The deprecation is just for the records. I guess we do not use the other methods.