Add form element value expectation descriptions where required
I did not adjust the DateTimePicker element as the sent format is
controlled by the useragent due to either the datetime-local input
type or a javascript datetimepicker. (which does not exist yet, though)
Mentioning the RFC 3339 format such a picker expects is very likely
going to lead to confusion on an end user's side otherwise.

refs #7947