lib: Add HttpNotFoundException
The HttpNotFoundException should be used for sending a HTTP 404 response w/ a custom message.

There's also Zend_Controller_Action_Exception but this exception will always show 'Page not found' because we
want to hide messages generated by Zend, like 'Action "foobar" does not exist and was not trapped in __call()'.

refs #6281