JS: Introduce getCSSPath
getCSSPath returns the CSS path to a given jQuery element. We have getDomPath which is not robust enough and
getElementByDomPath which makes no sense because getDomPath could already return a selector suitable for jQuery.
getCSSPath is meant to replace both of them.
The function is far from perfect, as its lacking class consideration and optimization.