IdoOpt: Optimize join performance for host list
The best possible join type in MySQL is `eq_ref` which is used when all parts of an index are used by the join and the index is a PRIMARY KEY or UNIQUE NOT NULL index.
The IDO schema already has some UNIQUE indices for joins but lacks NOT NULL in the column definitions.

refs #12732