IdoOpt: Optimize service_state performance
At the moment it's impossible for Web 2 queries which filter for host or service state to use indices because they
respect the virtual state PENDING. A host or service is PENDING if it has not been checked yet. Instead of calculating
the state on query execution, we store it for every record taking PENDING into consideration. This reduces database
load and enables MySQL to use possible indices.

refs #12732