Fix wrong unhandled services count in host views when restrictions are used
The query for fetching the unhandled services count utilises the hosts query as subquery.
Restrictions are applied to both the hosts query and the query for the unhandled services count.
This leads to wrong results since the restrictions are already in place for the unhandled services count because we're using the hosts query as subquery.

refs #2822