monitoring: Clone the filter for the grid links
Before, it was impossible to change an existing filter via the filter
editor in the host and service group views because of an exception about
wrong IDs.
The filter editor sets the IDs correctly. But since we're adding the
filter again to another URL, FilterChain::addFilter() overrides the IDs
because it is not working with a clone.