js: Don't bypass our cache when reloading CSS
The timestamp previously added, forces the browser to ignore
its own cache and thus it doesn't transmit the `If-None-Match`
header. We then re-calculate the the CSS possibly for no reason.
If there is a re-calculation indeed required it should be still
done now since the ETag transmitted by the browser is outdated.

refs #4340