monitoring/list/hosts: Don't show unhandled service counts
It has too much problems. It wasn't even visible since the
initial release, at least not if there were many results.
It being visible was more a case of chance than guarantee.
We now also identified a major issue with restrictions and
the way we can assemble queries here. In short, there are
too much technical difficulties that we deemed it not
worthwhile to keep.