[WIP] Allow modules to register *framework* params
I'm not yet convinced about this solution. The name
*framework parameter* isn't always correct. (i.e. page)

Then there's also an alternative available by using the
`rel` attribute of anchors with a custom value. If it's
set our JS preserves (or doesn't) the `_dashlet` param.
It cannot do that for every url that's loaded into the
same container. (e.g. stat badges in the footer of lists)