js: Don't trigger beforerender or rendered for discarded content
A listener for `rendered` should not need to check if the content
really changed. Though, if such a listener is not triggered anymore,
`beforerender` listeners also must not be triggered, as they might
assume that the content is really being updated and their accompanied
`rendered` listener is triggered. (e.g. input-enrichment.js)

This might be breaking change and any `Behavior.renderHook` implementation
needs to be checked against it. Potentially also in third party modules.
As if such an implementation updates the container on its own,
`beforerender` listeners only have access to the updated container after
this change, while they had access to the original beforehand.

`rendered` listeners should not be that much affected, as for them the
change results in the same behavior as if no update has ever been
scheduled for the container.

fixes #5056