Add 'Download QR Code' action link
So the user can download the QR code to backup if they lose access
to their device.

The download currently does only work in chrome and firefox, but
not in safari. The reason is, that the action link uses
`ipl\Web\Url` which adds '///' between the schema and the base
path. The safari browser can't handle this.

'data:image/png;base64,...' is converted to
'data:///image/png;base64...'.